requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alpaca-trade-api==2.3.0
altair==4.2.0
apache-airflow[postgres]
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
deprecation==2.1.0
dill==0.3.6
entrypoints==0.4
executing==1.0.0
extra-streamlit-components==0.1.56
fonttools==4.38.0
frozenlist==1.3.1
gitdb==4.0.9
GitPython==3.1.29
google-api-core==2.10.0
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
google==3.0.0
googleapis-common-protos==1.56.4
hydralit-components==1.0.10
idna==3.3
importlib-metadata==5.0.0
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
matplotlib==3.6.1
msgpack==1.0.3
multidict==6.0.2
multiprocess==0.70.14
numpy==1.23.3
packaging==21.3
pandas-ta==0.3.14b0
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
pre-commit==2.17.0
prompt-toolkit==3.0.31
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
PyJWT==2.6.0
PyYAML==6.0
Pygments==2.13.0
Pympler==1.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydeck==0.8.0b4
pyparsing==3.0.9
pyproject-toml==0.0.10
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
pytz-deprecation-shim==0.1.0.post0
pytz==2022.2.1
requests==2.28.1
rich==12.6.0
rsa==4.9
schedule==1.1.0
scheduler==0.8.2
scipy==1.9.1
semver==2.13.0
simplejson==3.18.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.5.0
stocksymbol==0.0.5
streamlit-aggrid==0.3.3
streamlit-authenticator==0.2.1
streamlit-extras==0.2.4
streamlit==1.19.0
streamlit-option-menu==0.3.2
ta==0.10.2
tablib==3.2.1
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
websocket-client==1.4.1
websockets==10.3
yarl==1.8.1
zipp==3.9.0
fastapi==0.95.0 
pydantic==1.10.7 
starlette==0.26.1
constant==0.0.4
html5lib==1.1 
webencodings==0.5.1
bcrypt==3.2.0
fastapi==0.63.0
passlib==1.7.4
pydantic==1.7.3
pytest==6.2.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
requests==2.25.1
SQLAlchemy==1.3.22
starlette==0.13.6
uvicorn==0.13.3
